chore(Controller): move Controller.kt file into controller package
