+++ src/main/kotlin/lotto/controller/Controller.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.controller
 
 import lotto.model.LottoMachine
 import lotto.model.LottoPrinter
